renamed folder structure
