+++ pom.xml
@@ -9,7 +9,7 @@
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>org.example</groupId>
-    <artifactId>Piggy-Projects</artifactId>
+    <artifactId>Project-Planner</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <name>Piggy-Projects</name>
     <description>Piggy-Projects</description>
