--- src/main/java/PiggyProjects/Controller/SubprojectController.java
@@ -1,11 +1,0 @@
-package PiggyProjects.Controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-
-public class SubprojectController {
-    public SubprojectController() {
-
-    }
-}
