--- src/main/java/PiggyProjects/Controller/UserController.java
@@ -1,4 +1,0 @@
-package PiggyProjects.Controller;
-
-public class UserController {
-}
