--- src/main/java/PiggyProjects/Repository/TaskRepository.java
@@ -1,4 +1,0 @@
-package PiggyProjects.Repository;
-
-public class TaskRepository {
-}
