--- src/main/java/PiggyProjects/Repository/UserRepository.java
@@ -1,4 +1,0 @@
-package PiggyProjects.Repository;
-
-public class UserRepository {
-}
