--- src/main/java/PiggyProjects/Service/TaskService.java
@@ -1,4 +1,0 @@
-package PiggyProjects.Service;
-
-public class TaskService {
-}
