--- src/main/java/PiggyProjects/Service/UserService.java
@@ -1,4 +1,0 @@
-package PiggyProjects.Service;
-
-public class UserService {
-}
