+++ src/main/java/ProjectPlanner/Controller/CompanyController.java
@@ -1,4 +1,4 @@
-package PiggyProjects.Controller;
+package ProjectPlanner.Controller;
 
 public class CompanyController {
     public CompanyController() {
