+++ src/main/java/ProjectPlanner/Controller/ProjectController.java
@@ -1,4 +1,4 @@
-package PiggyProjects.Controller;
+package ProjectPlanner.Controller;
 
 public class ProjectController {
     public ProjectController() {
