+++ src/main/java/ProjectPlanner/Controller/SubprojectController.java
@@ -1,0 +1,8 @@
+package ProjectPlanner.Controller;
+
+
+public class SubprojectController {
+    public SubprojectController() {
+
+    }
+}
