+++ src/main/java/ProjectPlanner/Controller/UserController.java
@@ -1,0 +1,4 @@
+package ProjectPlanner.Controller;
+
+public class UserController {
+}
