+++ src/main/java/ProjectPlanner/Model/Subproject.java
@@ -1,4 +1,4 @@
-package PiggyProjects.Model;
+package ProjectPlanner.Model;
 
 public class Subproject {
     private int projectID;
