+++ src/main/java/ProjectPlanner/Repository/CompanyRepository.java
@@ -1,4 +1,4 @@
-package PiggyProjects.Repository;
+package ProjectPlanner.Repository;
 
 public class CompanyRepository {
