+++ src/main/java/ProjectPlanner/Repository/ProjectRepository.java
@@ -1,4 +1,4 @@
-package PiggyProjects.Repository;
+package ProjectPlanner.Repository;
 
 public class ProjectRepository {
 }
