+++ src/main/java/ProjectPlanner/Repository/TaskRepository.java
@@ -1,0 +1,4 @@
+package ProjectPlanner.Repository;
+
+public class TaskRepository {
+}
