+++ src/main/java/ProjectPlanner/Repository/UserRepository.java
@@ -1,0 +1,4 @@
+package ProjectPlanner.Repository;
+
+public class UserRepository {
+}
