+++ src/main/java/ProjectPlanner/Service/TaskService.java
@@ -1,0 +1,4 @@
+package ProjectPlanner.Service;
+
+public class TaskService {
+}
