+++ src/main/java/ProjectPlanner/Service/UserService.java
@@ -1,0 +1,4 @@
+package ProjectPlanner.Service;
+
+public class UserService {
+}
