+++ src/test/java/org/example/piggyprojects/ProjectPlannerApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class PiggyProjectsApplicationTests {
+class ProjectPlannerApplicationTests {
 
     @Test
     void contextLoads() {
